+++ scripts/listenerui.py
@@ -7,8 +7,6 @@
 # WARNING! All changes made in this file will be lost!
 
 from PyQt5 import QtCore, QtGui, QtWidgets
-from PyQt5.QtCore import pyqtSlot
-from qnodelistener import QNodeL
 
 class Ui_listener(object):
     def setupUi(self, listener):
@@ -40,15 +38,6 @@ class Ui_listener(object):
         self.horizontalLayout_2.addWidget(self.logView)
         self.horizontalLayout.addLayout(self.horizontalLayout_2)
 
-        self.nodo = QNodeL(self.logView) #create qnode to run the Publisher
-        self.nodo.updateLog.connect(self.updateLogView) ##connect signal and slot of listview
-        self.nodo.rosShutdown.connect(self.close)
-
-        self.flag = False #flag to use one time
-
-        self.connectButton.clicked.connect(self.talk) ## connect publish button with the talker function
-        self.quitButton.clicked.connect(self.close)
-
         self.retranslateUi(listener)
         QtCore.QMetaObject.connectSlotsByName(listener)
 
@@ -59,20 +48,3 @@ class Ui_listener(object):
         self.connectButton.setText(_translate("listener", "Connect"))
         self.quitButton.setText(_translate("listener", "Quit"))
 
-    def talk(self):
-        if not self.flag:
-            self.flag = True
-            self.nodo.start()
-
-    @pyqtSlot()
-    def updateLogView(self):
-        self.nodo.text_window.scrollToBottom()
-
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    listener = QtWidgets.QWidget()
-    ui = Ui_listener()
-    ui.setupUi(listener)
-    listener.show()
-    sys.exit(app.exec_())
